requirements/requirements-dev.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.3
django-extensions==2.1.4
django-filter==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djoser==1.3.1
idna==2.7
Markdown==3.0.1
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.20.1
six==1.11.0
toml==0.10.0
urllib3==1.24.1
